refactor: reorganize upload  classes PE-1178
