+++ lib/blocs/upload/models/upload_file.dart
