+++ lib/blocs/upload/models/upload_plan.dart
@@ -1,16 +1,16 @@
-import 'package:ardrive/blocs/upload/bundle_upload_handle.dart';
+import 'package:ardrive/blocs/upload/upload_handles/bundle_upload_handle.dart';
 import 'package:ardrive/utils/bundles/next_fit_bundle_packer.dart';
 
-import 'data_item_upload_handle.dart';
-import 'file_upload_handle.dart';
+import '../upload_handles/file_data_item_upload_handle.dart';
+import '../upload_handles/file_v2_upload_handle.dart';
 
 final bundleSizeLimit = 503316480;
 final maxBundleDataItemCount = 500;
 final maxFilesPerBundle = maxBundleDataItemCount ~/ 2;
 
 class UploadPlan {
-  /// A map of [FileUploadHandle]s keyed by their respective file's id.
-  late Map<String, FileUploadHandle> v2FileUploadHandles;
+  /// A map of [FileV2UploadHandle]s keyed by their respective file's id.
+  late Map<String, FileV2UploadHandle> v2FileUploadHandles;
 
   final List<BundleUploadHandle> bundleUploadHandles = [];
 
@@ -19,8 +19,8 @@ class UploadPlan {
   });
 
   static Future<UploadPlan> create({
-    required Map<String, FileUploadHandle> v2FileUploadHandles,
-    required Map<String, DataItemUploadHandle> dataItemUploadHandles,
+    required Map<String, FileV2UploadHandle> v2FileUploadHandles,
+    required Map<String, FileDataItemUploadHandle> dataItemUploadHandles,
   }) async {
     final bundle = UploadPlan._create(
       v2FileUploadHandles: v2FileUploadHandles,
@@ -30,10 +30,10 @@ class UploadPlan {
   }
 
   Future<void> createBundleHandlesFromDataItemHandles(
-    Map<String, DataItemUploadHandle> dataItemUploadHandles,
+    Map<String, FileDataItemUploadHandle> dataItemUploadHandles,
   ) async {
     // NOTE: Using maxFilesPerBundle since FileUploadHandles have 2 data items
-    final bundleItems = await NextFitBundlePacker<DataItemUploadHandle>(
+    final bundleItems = await NextFitBundlePacker<FileDataItemUploadHandle>(
       maxBundleSize: bundleSizeLimit,
       maxDataItemCount: maxFilesPerBundle,
     ).packItems(dataItemUploadHandles.values.toList());
