+++ lib/blocs/upload/models/web_folder.dart
@@ -1,13 +1,10 @@
-import 'package:ardrive/blocs/upload/web_file.dart';
-
 class WebFolder {
   final String name;
   final String parentFolderPath;
 
   String id;
-  List<WebFolder> subFolders = [];
-  List<WebFile> files = [];
-
+  late String parentFolderId;
+  late String path;
   WebFolder({
     required this.name,
     required this.id,
