+++ lib/pages/drive_detail/components/drive_file_drop_zone.dart
@@ -1,5 +1,5 @@
 import 'package:ardrive/blocs/blocs.dart';
-import 'package:ardrive/blocs/upload/web_file.dart';
+import 'package:ardrive/blocs/upload/models/web_file.dart';
 import 'package:ardrive/components/upload_form.dart';
 import 'package:ardrive/models/daos/drive_dao/drive_dao.dart';
 import 'package:ardrive/services/services.dart';
